Fixed some issues with database interface method signatures. Added location to remote reader. Added code to convert a RemoteRead to a ChipRead for saving to the database. Renamed the RemoteReaders window to RemoteReadersWindow.

+++ ChronoKeep/Network/Remote/RemoteHandlers.cs
@@ -50,7 +50,7 @@ namespace Chronokeep.Network.Remote
             throw new APIException(content);
         }
 
-        public static async Task<GetReadsResponse> GetReads(APIObject api, string reader, int start, int end)
+        public static async Task<GetReadsResponse> GetReads(APIObject api, string reader, long start, long end)
         {
             string content;
             Log.D("Network.Remote.RemoteHandlers", "Getting reads.");
